feat(cb2-6538): rename file
